Dynamic Piece Size
It tries to calculate the log2 of the torrent size in order to dynamically pick the best 2**x piece size.

If that doesn't give a correct answer, it uses the table to choose the piece size.

I updated the table, basically updating the piece exponent +2 from its original configuration.
This table gives the range of piece sizes from: 512 - 3700. While the original configuration used 4096 - 8192 pieces.